Conflicts:
	site_scons/site_tools/integrate_bazel.py
	src/mongo/db/SConscript
	src/mongo/db/auth/SConscript
	src/mongo/db/auth/authorization_manager_impl.cpp
	src/mongo/db/auth/authorization_manager_impl.h
	src/mongo/db/catalog/SConscript
	src/mongo/db/commands/SConscript
	src/mongo/db/concurrency/BUILD.bazel
	src/mongo/db/exec/sbe/BUILD.bazel
	src/mongo/db/ftdc/SConscript
	src/mongo/db/mongod_main.cpp
	src/mongo/db/op_observer/SConscript
	src/mongo/db/repl/SConscript
	src/mongo/db/serverless/SConscript
	src/mongo/db/storage/SConscript
	src/mongo/idl/SConscript
	src/mongo/s/commands/SConscript

+++ evergreen/generate_evergreen_bazelrc.sh
@@ -9,8 +9,14 @@ set -o verbose
 # Use the Evergreen temp directory to avoid filling up the disk.
 mkdir -p $TMPDIR
 if [[ "$OSTYPE" == "cygwin" ]] || [[ "$OSTYPE" == "win32" ]]; then
+  mkdir -p Z:/bazel_tmp
+  touch Z:/bazel_tmp/mci_path
   # TODO(SERVER-94605): remove when Windows temp directory is cleared between task runs
-  rm -rf Z:/bazel_tmp/* || true
+  if [[ "$PWD" != "$(cat Z:/bazel_tmp/mci_path)" ]]; then
+    echo "Clearing bazel output root from previous task mci '$(cat Z:/bazel_tmp/mci_path)'"
+    rm -rf Z:/bazel_tmp/* || true
+    echo $PWD > Z:/bazel_tmp/mci_path
+  fi
 
   # Z:/ path is necessary to avoid running into MSVC's file length limit,
   # see https://jira.mongodb.org/browse/DEVPROD-11126
@@ -29,12 +35,12 @@ if bazel_rbe_supported; then
 
   uri="https://spruce.mongodb.com/task/${task_id:?}?execution=${execution:?}"
 
-  echo "build --tls_client_certificate=./engflow.cert" >> .bazelrc.evergreen
-  echo "build --tls_client_key=./engflow.key" >> .bazelrc.evergreen
-  echo "build --bes_keywords=engflow:CiCdPipelineName=${build_variant:?}" >> .bazelrc.evergreen
-  echo "build --bes_keywords=engflow:CiCdJobName=${task_name:?}" >> .bazelrc.evergreen
-  echo "build --bes_keywords=engflow:CiCdUri=${uri:?}" >> .bazelrc.evergreen
-  echo "build --bes_keywords=evg:project=${project:?}" >> .bazelrc.evergreen
-  echo "build --remote_upload_local_results=True" >> .bazelrc.evergreen
-  echo "build --workspace_status_command=./evergreen/engflow_workspace_status.sh" >> .bazelrc.evergreen
+  echo "common --tls_client_certificate=./engflow.cert" >> .bazelrc.evergreen
+  echo "common --tls_client_key=./engflow.key" >> .bazelrc.evergreen
+  echo "common --bes_keywords=engflow:CiCdPipelineName=${build_variant:?}" >> .bazelrc.evergreen
+  echo "common --bes_keywords=engflow:CiCdJobName=${task_name:?}" >> .bazelrc.evergreen
+  echo "common --bes_keywords=engflow:CiCdUri=${uri:?}" >> .bazelrc.evergreen
+  echo "common --bes_keywords=evg:project=${project:?}" >> .bazelrc.evergreen
+  echo "common --remote_upload_local_results=True" >> .bazelrc.evergreen
+  echo "common --workspace_status_command=./evergreen/engflow_workspace_status.sh" >> .bazelrc.evergreen
 fi
